requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.346
langchain-groq==0.1.0
langchain-core==0.1.16
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
streamlit>=1.28.0
requests>=2.31.0